Continue implementation
