+++ src/modules/src/Eryph.Modules.VmHostAgent/UpdateConfigDriveCommandHandler.cs
@@ -66,7 +66,7 @@ internal class UpdateConfigDriveCommandHandler :
                 .Bind(o => o.ToEither(Error.New("Could not find storage settings for VM.")).ToAsync())
             let genepoolReader = new LocalGenepoolReader(vmHostAgentConfig)
             from mergedConfig in fodderConfig.BreedAndFeed(genepoolReader, metadata.ParentConfig)
-                .Map(c => c.AppendSystemVariables(metadata))
+                .Map(c => c.FeedSystemVariables(metadata))
                 .ToAsync()
             from substitutedConfig in CatletConfigVariableSubstitutions.SubstituteVariables(mergedConfig)
                 .ToEither()
